Create 11.3.Building Promises.all.js
As we saw, given an array of promises, Promise.all returns a promise that waits for all of the promises in the array to finish. It then succeeds, yielding an array of result values. If a promise in the array fails, the promise returned by all fails too, passing on the failure reason from the failing promise.

Implement something like this yourself as a regular function called Promise_all.